chore(mixins): Updated structure
