+++ src/mixins/data-workflows/collection/configuration-collection-mixin.ts
@@ -1,3 +1,5 @@
+// TODO: Create common mixin & import components
+
 import { Component, Vue } from 'vue-property-decorator';
 import { CONFIGURATIONS } from '@/constants/data-workflows/status';
