--- src/mixins/data-workflows/configuration-doc-mixin.ts
@@ -1,10 +1,0 @@
-import { Component } from 'vue-property-decorator';
-import ItemMixin from '@/mixins/data-workflows/item-mixin';
-import { CONFIGURATIONS } from '@/constants/data-workflows/status';
-
-@Component
-export default class ConfigurationDocMixin extends ItemMixin {
-	get type() {
-		return CONFIGURATIONS;
-	}
-}
