+++ src/mixins/data-workflows/doc/item-mixin.ts
@@ -4,7 +4,7 @@
  */
 
 import { Component, Vue, Watch } from 'vue-property-decorator';
-import { AnyObject, DataWorkflowsType, ConfigurationProps } from '@/types';
+import { AnyObject, DataWorkflowsType } from '@/types';
 import { IPluginState } from 'vuex-easy-firestore/types/declarations';
 import DataManagementHeader from '@/components/data-workflows/common/DataManagementHeader.vue';
 import ItemComponent from '@/components/data-workflows/common/item/ItemComponent.vue';
@@ -56,14 +56,6 @@ export default class ItemMixin extends Vue {
 		await store.dispatch(`${this.moduleName}/fetchById`, this.itemId);
 	}
 
-	get updateInformation() {
-		// TODO: Add UserSocialInfo
-		return {
-			update_date: this.item.update_date || this.item.updated_date,
-			updated_by: this.item.updated_by
-		};
-	}
-
 	get itemId(): string {
 		return this.$route.params.id ? this.$route.params.id : ' ';
 	}
@@ -71,18 +63,5 @@ export default class ItemMixin extends Vue {
 	// Overridden by component
 	get type(): DataWorkflowsType {
 		return null;
-	}
-
-	get configurationProps(): ConfigurationProps {
-		return {
-			archivedConfsModuleName: this.archivedConfsModuleName,
-			docId: this.itemId,
-			isLoading: this.isLoading,
-			isNotFound: this.isNotFound,
-			moduleName: this.moduleName,
-			tabsItems: this.itemTabsItems,
-			type: this.type!,
-			updateInformation: this.updateInformation
-		};
 	}
 }
