+++ src/mixins/data-workflows/doc/run-doc-mixin.ts
@@ -1,10 +1,11 @@
 import { Component } from 'vue-property-decorator';
-import ItemMixin from '@/mixins/data-workflows/item-mixin';
+import { DataWorkflowsType } from '@/types';
+import ItemMixin from '@/mixins/data-workflows/doc/item-mixin';
 import { RUNS } from '@/constants/data-workflows/status';
 
 @Component
 export default class RunDocMixin extends ItemMixin {
-	get type() {
+	get type(): DataWorkflowsType {
 		return RUNS;
 	}
 }
