+++ src/views/data-workflows/storage-to-tables/runs/ListingView.vue
@@ -30,7 +30,7 @@ import DataManagementHeader from '@/components/data-workflows/common/DataManagem
 import ListingComponent from '@/components/data-workflows/common/listing/ListingComponent.vue';
 
 import HeaderInfosMixin from '../header-infos';
-import RunCollectionMixin from '@/mixins/data-workflows/run-collection-mixin';
+import RunCollectionMixin from '@/mixins/data-workflows/collection/run-collection-mixin';
 
 import { STORAGE_TO_TABLES_RUNS_ITEM } from '@/constants/router/routes-names';
